feat/rest-api: finish line chart for suatu pasar

+++ src/main/java/com/menyala/sipm/rest/dto/response/LineTransaksiSuatuPasarResponseDTO.java
@@ -9,6 +9,6 @@ import java.util.List;
 @NoArgsConstructor
 public class LineTransaksiSuatuPasarResponseDTO {
     private String namaPasar;
-    private List<Integer> data;
+    private List<Long> data;
     private List<String> labels;
 }
